allow tasks to receive json and pass csv parsing options to readers
Passing various parameters individually to tasks as command line args
was getting really messy and restrictive. Instead, the entire json
message received from the queue on the main process is now passed to
the task as a serialized argument.

This paves the way to optionally pass more parsing options in an import
task.