update models to require less boilerplate and more easily evolve
There's a tangle of calls back and forth that's only gotten worse over
time. The set of changes I'm looking to make would make it worser still if
not addressed. This may be a fool's errand at this point, but I'm gonna try.

The basic idea with this is to let the neo4j node properties do the work of
populating model object properties and to avoid trying to populate various
related entities all at once. By giving the models themselves access to doing
their own neo4j queries, a lot of the complexity of accessing related
information can be abstracted away.