can save on subclasses of the base model and moved connections code
You can now call save on any object that extends the Base model. This can
be overridden in subclasses to provide more custom save behavior. The base
behaviour uses the label and uuid to find the neo4j node to save to and
uses a saveProperties array class variable to determine which properties
to actually try to save. Any properties in that list that have been set
on the object will be saved.