populate the current user from a graphql query once we login
This gets the site working again after changes to the user model, but we
could likely go further fairly easily and remove some of the stranger
workarounds by storing the 'currentOrg' in redis and returning that along
with the 'currentUser'. This would allow other components to operate
independently and remove the need to pass so much around in the
navigation context.