fix package dependencies resulting in weird graphql error
Props to Alecia for tracking down the discussion about:
"Encountered a sub-selection on the query, but the store
doesn't have an object reference.":

https://github.com/apollographql/apollo-client/issues/4125

We weren't directly using apollo-cache-inmemory, which was
the workaround suggested here. But we were using older
versions of apollo-boost, from which we were getting the
InMemoryCache object. They have probably updated since.