change file storage to handle multiple versions of csv files
This allows us to store the original file alongside normalized files in
a way that either can be easily retrieved. This makes it easier for the
import_csv task to deal with variations of CSV files while being able to
provide a single format expectation post-import.

A migration script is also included to translate the old stored files
to new locations.