fix subscription to dataset updates
The key issue was that 'datasets' was undefined and that error was only
showing up later in the chain. It seems that there's something on the client
side that can cause the momentary state where loading is true, error is
undefined, and yet data.dataset is also undefined. Once we guarded against
this possibility, the subscription magically started working again.

However, it couldn't have helped that we were also (probably) subscribing
over and over again, on every render. With better understanding of the
React lifecycles since the original code was written, and re-exploring the
apollo graphql docs, it looked best to split into multiple components and
subscribe in a lifecycle hook.