fix subscription when dataset is brand new
Subscriptions were actually working already (after the previous fix) when
the dataset had already been initialized with data, and either another
upload was made or the dataset was regenerated. However, when the dataset
was just created, the subscription would run into an issue with state
variables incorrectly referenced where the subscription handler was moved
to. Short term, passing in a function to handle the status potentially
reported (and set on the state) allows 'this' and 'this.state' to be
referenced in the correct context. Further down the road, we'll likely
want to investigate doing this without HoCs to potentially simplify the
logic. Still, hooks should ultimately make this a piece of cake.