use common csv output functionality to make sure all important files get written
Some of the issues with transformations not seeming to work at the end had
to do with the sample.csv file not existing under the new CSV paths when a
new transformation was done on a new dataset. Now the same functionality is
used from import_csv in order to write out all important files, including
sample.csv.