really, really basic integration test runner
Still to do on this: actually generate the test csv files. Right now,
these reference files that we've used testing by hand. But ideally we'd
want the integration test to just generate the sample files it needs.
However, this still provides some use now to test just the storage
functionality without all the in-between pieces, with minimal setup.