fix upload of large files
This is a small one line fix with a big comment added to reflect the amount
of grief the problem gives and how mysterious it seemed. In hindsight, the
problem was pointed out specifically in the graphql-upload tips, and
correct code was in the package's examples. However, it was easy to miss
what was actually so important in the example, so the comment here was
added for anyone in the future who might play with this area of the code.