use amqp-connection-manager to keep queue connections alive
Made changes to the queue library so that it uses the connection manager
to create connections. The setup is very similar, but the connection manager's
slightly more verbose setup allows it to manage keeping the queue alive,
which is especially important due to recommended keepalive timeouts
occasionally disconnecting it.

While there, I also made some other changes to further encapsulate the
queue connection, making it easier to shut down when the site shuts down.