fix auto-reload and update bin script to allow pdf/epub generation
The "fix" for auto-reload is basically to go back to the defaults, as at
least the live-reload custom port doesn't really work:

https://github.com/GitbookIO/plugin-livereload/pull/1

The bin/manual script also now needs "start" to start the server, as it's
now multi-purpose, also handling stopping the server and generating pdfs
and epubs.

Finally, the server shouldn't constantly re-install gitbook with every
run now.