puneo4j Dockerfile: chown plugins directory completely
While the contents of the neo4j plugins directory in the Dockerfile
were being chowned appropriately, the actual plugins directory itself
wasn't.

Since the neo4j container dies on startup if it can't write there,
this is sub-optimal.

It's still unclear why this works for some people in some cases.