use a provided map of dataset names to locations in generate script
This moves more logic for knowing which underlying files should be modified
into the server, where most of the other business logic resides. Remaining
work will involve doing the same for the storage output path. With that in
place, we should be able to begin work in the server that further abstracts
Datasets into at least a couple of high level subsets, representing
sturctured and unstructured data.