oops... we actually don't want to raise this error
Previously, we were doing some refactoring and came across the generally
big no-no of semi-silently swallowing an error. At least that's what we
thought. Turns out, we can't just tell everything to blow up at this point
because we still want to send a message back to the server letting it know
what happened. And we actually send the error that occurred along at that
point. So this raise needs to get removed.