standardize our 'error' status
In our register_transformation task, we were returning 'error' if something
went wrong, and in generate, we were returning 'failure'. The only real
problem with this was that in some cases the problem would percolate up
enough for the interface to display it, and in other cases it wouldn't.

We could look for both, but simply using one or the other means we only
have to look for the one thing. There doesn't seem to be any good reason to
differentiate between 'failure' and 'error', so I flipped a coin and
chose 'error'.