accessor functions for Organization and Dataset
We didn't have a function to get an organization by its unique name,
so I added it. Then I used this in a Dataset function, getByFullName, to
allow using the 'org:dataset' style fully qualified name to retrieve a
dataset.

Right now, this makes things useful in passing information back from the
worker processes. They often figure out a dataset through one of the magic
dataset_input function calls, and they really only care about the path to
the physical dataset, which we're now passing in. So we could avoid
the need to pass names by also passing in IDs, but we'd only really be doing
that just to pass them back again, which seems rather silly.

Add to this that further refinement of the process (adr in progress) will
likely eliminate many of these problems, and it comes down to the question
of which temporary fix will be less bothersome to remove. This will be both
less bothersome and, conceivably, still of use in other areas.