give responsibility back to the server for updating columns
This dumbs down the generate task further. Previously, each transformation
it went through, it would update the columns in the database, in case they
changed. This sort of thing keeps the roles of the worker tasks and the
server confusing. Not only is the worker responsible for crunching the data,
but it also has to know update the database in the exact way the server
expects it, meaning both have to know a lot about the underlying data model.

So instead, we're gathering up the changed columns and sending them back in
the update message in a way the server can use to update its data model.

There's one hitch here. Potentially many datasets will have column changes,
as all the transformations in the chain are run. But we're only sending a
single queue message for each task completing. We were (temporarily)
handling queue updates on the affected dataset. But in this case it would be
really weird to make one dataset responsible for coordinating changes to
the metadata of a bunch of other datasets. So it made sense to have something
outside of the existing domain objects handle this.

Meanwhile, there are many painpoints in this back-and-forth that I'm
convinced would benefit from a Task node and associated logic that focuses
on the coordination and tracking of work. It sounds fancier than I envision
it, and an ADR is pending. But I think it sets us up for a lot of good things
including finally being able to distribute a long chain of transformations.

Anyway, it seemed like just a bit too much of a diversion to try and do that
during this story, so I'm mainly just using it as a place to put the
functions to coordinate the dataset updates, ultimately expecting that the
code will wind up in its own Task class or subclass.