actually pass back datasetColumnUpdates for use
Without this, the changes in the 8001288e85c0fcdd309c4a0228cab645ffed4916
commit won't work. They shouldn't actually break anything, but the old
method of the worker updating column metadata would be used and relied upon.

This change both removes that old update method and sends the necessary
information back to the server for it to use its update method.