fix downloads once again
Yesterday, we changed some of the information being passed to the generate
task. The download task is (unfortunately) at the moment tied pretty closely
to the generate task. There's no real smarts going on, so the download just
always assumes the datasets may not have been generated and makes sure to
generate them first, meaning it needs all that same information.

Once we start to be able to process transformation chains in pieces, we
can likely remove this duplicate responsibility from the download task.