Fix some weird regressions with uploading
I'm actually not quite sure how this seemed to work before, as the
previous version was passing the first row returned without getting the
actual dataset object (d at the time). Previous to that, we'd been
explicitly loading by ID, but I thought I could be smart and avoid a
database query given the handy dandy ModelFactory loading. Nope.
Even after correctly getting the dataset contained within the returned
row, that doesn't get a uuid until the transaction is saved.

It would be nice to keep this bit of weirdness at the database layer. It
makes me wonder if there's an ability within a Neo4J query to tell it
to commit a transaction and then run another query before returning.
That way the database layer could be dealing with its own quirks while
the server layer could remain delightfully oblivious. Until that magical
time, I've updated the warning to hopefully explain better why we have
to do things the way we are and avoid another gotcha.