use the ModelFactory so subclasses can decide on how they should be "downloaded"
Classic issue created by the refactoring. Pretty much anywhere there's a
Dataset.get, we should be using ModelFactory.get instead. This will
properly decide on the actual JS class for the object. In this case, we
needed to do that because we specifically output our structured datasets
as csv files.

With this, the integration tests created earlier now output expected
results again, and we can get to refining the two different styles
of Dataset uploads.