only set type if it's not there to begin with
We want to be careful when setting values in the constructors that exist
as database properties. If we're using a dataset subclass to create a
new one, this is expected behaviour (although we'll still likely be
getting a type specified from somewhere above), but if we're loading
from the database or specifying a set of properties (which get set
within the Base superclass), it would be unexpected to override them
immediately.