deal with a remaining cyclic dependency
There's a bit of a smell in having to import the ModelFactory to make
sure everything's properly registered, and certainly room for
improvement. We should still be on the lookout for better solutions for
this sort of thing. Even so, it is an improvement over the way things
were up to this point.