say hello to first class document support!
With this change, we can call:

dataset.upload('document', '..path', type='document')

with a modified version of our python API (in other words, we can
trigger a GraphQL request where we specify a type and then another
GraphQL request to upload to that typed dataset) and actually have the
file stored without renaming it to look like a .csv. In fact, we use the
file extension from the original file (which we may want to consider
doing even with the .csv datasets, which are really meant to be just
structured data) when storing in object storage (as 'original.pdf' for a
pdf, for example). We still use a normalized naming scheme, while
storing the original filename as a separate property of our dataset.