Don't try to import documents.
I could see an argument for maybe not even having to do a real
"transformation" to get raw text for a dataset, and this could then be
the import step. But I'm not sure we want to really do that.

If there's time, I might see what that would look like, but currently
it's not critical to the story, and we haven't really talked about the
consequences of putting more on that import task that we would otherwise
do in transformations. For example, what if the import isn't quite how
they wanted and a transformation was needed anyway? Then we're doing
some useless steps. If they're not too costly, that may still be fine
(and results can still be used for previews).