allow a Transformation to be defined as an ALIAS_OF another
In this case, we follow the ALIAS_OF connection to the actual
Transformation we want to use. This will mean we can leave the existing
Transformation nodes where they are, which are still useful in
maintaining individual data transformation paths, but they can
ultimately be resolved by another more universal Transformation.

The next steps will be providing a way for a user to wire the inputs
differently and then enter these more independent transformations
without having to define a dataset.