if set, use alias on input edge when mapping storage
Yay graphs! Our current structure actually makes for a pretty elegant
solution to untie dataset inputs from the scripts themselves.

Every computed dataset is connected by an OUTPUT edge of a
Transformation. There will also be INPUT edges going into that
Transformation. If this is going to be just a placeholder Transformation
node that references one of our reusable Transformations, then we can
use those INPUT edges to define *how* the input Dataset is being used
in the Transformation. Specifically, since we have to use a placeholder
name within the transformation script, but we actually want a
particular Dataset input to represent that placeholder, we just use the
alias as a key for mapping the stored dataset file instead of the actual
Dataset name. It results in a surprisingly small amount of change to
support reusable Transformations and is also backwards compatible with
the way we have been defining them (as one offs), which may still be
desirable in many situations.