provide common functions for creating nodes and attaching relations
We're creating nodes largely the same way in a bunch of places but with
custom queries. Instead of furthering that, I'm creating a hopefully
generic enough function that can be used to create and set properties
for any Neo4J node.

Similarly, there's a lot of repetition in establishing relationships.
The real place where there still might be a lot of legitimate variation
is in the queries exploiting the relationships for various bits of
functionality. But where when we're establishing those relationships,
we're usually doing it one at a time, so something simple like
saveRelation could cut down on a lot of custom code.

Part of me worries that, in doing things like this, we're creating a lot
of queries for simple things that could be done in one query. But there
are better ways to make that more efficient when we have to, the most
obvious being to start grouping multiple queries for creating initial
states of nodes and relations into transactions. This will no doubt
require some modification to these functions, but they seem like a step
in the right direction.