eliminate some duplicate logic on Transformation creation
There will probably be some more changes with codeless transformations
but right now, transformations are pretty meaningless without code,
which is stored as a separate file, and this was being left to anything
trying to create a transformation to figure out the right way to fully
initialize it. Now that logic is more central to the Transformation
class and its static create function.