factory functions to get by uuid and name
It's becoming increasingly useful to have the ModelFactory handle
fetching and instantiation of objects. These mirror functions in the
Base class, and can generally be used in their place. One exception to
note is the getByName function, which really needs to know the Neo4J
label and owner to reduce the scope.