add more comprehensive reference types and ways to resolve them
Transformations and datasets are often referred to by name or id. We
want to move towards uuids instead of id, but we'll also always want to
be able to resolve a name.

GraphQL allows us to provide for these different ways of identifying
these references by defining specific input types. We need to know how
to resolve them on the other side, so utility functions are also being
added to do that.