make saveRelation a little more handy
Previously, saveRelation would only relate a node to the current node.
This covers many cases, but there are others where we need to specify
both sides of the relationship and it makes sense to have something
that's not directly a part of the relationship coordinating things.

The main use case for this at the moment is the new reusable
transformations, where we need to do more wiring together of input/
output datasets up front.