remove checking for this.path before retrieving from storage
We've long since moved from the single path on a dataset model, towards
deriving the paths based on uuids and specific needs of the dataset type
(for example, csv datasets compute imported and sample versions).

We should probably do a larger search through the codebase and cut this
particular attribute off at the root.