add alias to the INPUT relations when using reusable transformations
This includes some changes to saveRelation on the Base class, currently
only used by other changes in this branch. The changes allow us to set
attributes on a saved relation itself.

As well, note that placeholder has been changed to alias to mirror other
code and not mix metaphors too much.