return the template's code block if referencing a template
When our transformation is really just a placeholder for another, we'll
return that other transformation's code.

We won't allow setting the template's code through a transformation
referencing it. This would be undesired behaviour in most cases. In
order to differentiate in the web ui or other places that need to,
we'll provide an indication of whether or not the directly attached
transformation is playing the role of a referernce.

In a more user oriented perspective, that means that the transformation
code they're looking at is part of a reusable transformation.