resolver debugging helper and removing cruft
One of the debug lines was overly complicated, so I moved some of that
code to the utility module being used by resolvers. Looks like I'd also
missed a placeholder => alias conversion from the 293a782b95 commit.

Finally, removing the passing of context.user. We only need it when
creating the original types of transformations, and even then, we only
need it for the register_transformation queue task. As we continue to
make the worker queue dumber, we should need to pass less of this sort
of information.