unbreak the web ui for reusable transformations
When a dataset is computed with an input transformation that references
a reusable transformation, the code for the referenced transformation
will show up in the editor, but it will not be editable. The save button will also be removed.

This is all made possible by passing back a 'virtual' boolean property
when querying a Transformation. This is enough, as even if hacked
around, you cannot save directly to a reusable transformation by saving
an input transformation on a dataset that references the reusable one.