pass organization in for resolving reusable transformations
When referring to a reusable transformation by name, we need to know the
organization. So we need to pass this in as well.

I also updated the organization find function to do the check that the
user is actually part of the organization, as this is needed for
any action where we're doing it under the organization.