fix occasional error when saving reusable transformations
This was getting triggered by the directive that checks access rights.
It was a weird one because, while everything in our resolver definition
seemed to work, and the right objects got created (and sometimes, the
right timing would mean no error at all), we would still return an error
to the client using the API. Turns out it was one of those pesky async
issues. In particular, though 'return' is fine from an async function
when calling another async function, without first doing an 'await', if
you actually aren't returning anything, you still need to 'await', even
if it's the last line in an async function. It's not the property of
being the last line. It's the return that makes sure to make the promise
chain happy. You still need to make sure to await when calling any of
these async functions and using results after in the same function.

There are quite a few places where we're not doing this with safeQuery.
Most are probably harmless, but next commit I'll be fixing a bunch of
them. We should probably look out for others. safeQuery just happens to
be an async we use in a lot of places and often don't need results from.