fix not awaiting async functions when we need to
See commit b462df2e9 for further explanation. Basically, if we're not
returning in the last line of an async function, and that last line is
calling an async function, we still have to await it. When we are
returning, we don't have to await (and that seems the recommended style
still).