add p-wait-for for things we have to wait for
Turns out in this brave new world of asyncronous services, you can't
always count of things being ready when you try to access them. Duh.

Most of the time, it's fine to send off messages on the queue or files
to object storage and they'll finish when they finish (and, in fact,
they'll let you know when they're done). But sometimes, you're impatient
and want to execute things really fast in a notebook. And then you end
up having to make sure you wait for certain crucial things to finish
before going on to the next thing.

See further commits for this in action.