fix race condition with transformation registration
Because the old school transformations need to be run through the
register_transformation task in order to figure out the actual inputs,
we really have to wait for that before we can use the transformation
chain to generate our datasets. This is often not noticeable when
running by hand through the web UI, but when driving it through a
notebook, it can result in some weird errors.

This addresses the problem by first attaching a state of 'registering'
to an old school transformation that only gets converted to 'ready' when
the registration task completes and messages back. Then we make sure
when computing the transformation chain to wait for the ready state
(still with a reasonable timeout because we don't want to wait forever
and would rather trigger an error if things are that bad). Finally, new
reusable transformations can start as 'ready' because we specify their
inputs when we define a reference to them.