remove NOT EXISTS check that can break transformation chains
This was a strange odyssey. In our query to figure out all the
transformations to run when generating a dataset, there was a part in
the middle that was inexplicably checking:

AND NOT EXISTS(t.error)

effectively removing any transformations that were unlucky enough to get
tagged with an error. This could result in some weird behaviour with the
script editor when you tried to generate using code that had an error,
fixed it, then tried again. You could usually button mash your way to
victory, but not without being very confused as to what actually is
wrong.

So, the easy thing to do was remove the offending NOT EXISTS. After all,
the errors would still get passed up. And if you had transformations
that still had errors in them, they just wouldn't run and keep reporting
those errors, which is kind of what you want.

However, why had it been added in the first place? It looks like I had
added it with questionable logic, to ensure you couldn't run
transformations on datasets you didn't have access to through one of
the organizations you're  a member of.

One of the major errors the python-workers could throw was if you tried
using an org:dataset_name reference with an org you were not a member
of. It would throw that error, and then you would know that that
transformation couldn't be run, due to not having permission. So then I
guess I figured all errors should keep you from having a transformation
in the chain because some errors were authorization ones?

Better to explicitly check that sort of thing, which is easier now with
our growing domain layer. I've added a canTransform function and
threaded passing a user down to the datasetStorageMap function, which is
the one gathering up all the input and output datasets for a chain of
transformations. It seemed as sensible a place as any to do an extra
check that the user running a transformation should really be able to
do it. As explained in a comment in the datasetStorageMap function,
we may have to revisit the more restrictive nature of canTransform,
which effectively stops a whole transformation chain if one dataset in
it is not accessible by the user.