populate the repl with all models in models.js
This is instead of loading and attaching each individually. Note that I
ran into a bit of a hiccup while doing this due to the User class's use
of bcrypt, which needs different binaries between macos and linux.
Previously, we'd never loaded User up in the repl, so hadn't run into
the problem here. I don't know why the server still seemed to be
working. In any event, if you run into the problem, you need to log into
your server container and do an npm install. While the repl seems better
at drawing the problem out, it has nothing to do with the actual
problem.