simplify interface for retrieving related nodes
First change: explicitly ask for the label, not the model class. We're
getting more cases where the actual object class we want to end up using
is based on a type attribute, and for this reason, we don't really know
what class we'll want to use until we load it. There's a derive function
in ModelFactory that helps figure out what class to actually use, given
a returned chunk of json representing a neo4j object.

Second: Previously, we were requiring a name to be passed in for the
reference to the related node(s) we're trying to retrieve. We actually
don't care what the reference is. There are only one reason you may
need it: to retrieve the right json object out of a result row. We've
since standardized these queries enough that we can pick any single name
within the functions that create the relation query and then just make
sure to use it when we're pulling the node out –there's no reason to
expose that detail to the outside.

When we save relations, we're still requiring a potential reference name
in case we have to attach properties to the relationship itself. This
could (and probably should) be eliminated with a little creative coding.