derive unique match to get the right model class
As we get more model subclasses, it's more important to make sure we're
creating the right class to work with them, based off the label and
type. Now getByUniqueMatch will use ModelFactory.derive, throwing an
error if you're trying to do something silly like trying to use a
totally unrelated model class to fetch an object by uuid. We may decide
later to relax this, as we get more confident with the ability to
properly derive the right object class, but for now, a little extra
checking will prevent surprises. If we do relax it, these sorts of
functions should probably end up somewhere else anyway.