complete basic Task chain and TransformTask logic
This further defines some basic bookkeeping that the Task base class
carries out and also how its subclasses should interact with it. It
also further defines the TransformTask subclass closer to where it will
actually do something.

datasetStorageMap is being temporarily duplicated. This version is
changed slightly because it no longer has to deal with an array of
transformation ids. It's also been updated to grab the uuid instead.
When this is fully rolled out, other code should no longer need to call
this function directly. It will only have to make a TransformationTask.
Currently the only thing outside of generating a dataset that needs
these mappings is download functionality, and that's because it has had
to also do the generation up to this point (due to not being able to
deal with coordinating work). Under the new Task system, we'll just add
a DownloadTask as the next after a chain of TransformTasks.