have runTransformation create and kick off a chain of tasks
A temporary duplication of parentTransformations now exists as
transformationChain. It can be slightly simpler, as it just has to
return the transformations in the correct order. Each of those will be
turned into a TransformTask connected to the transformation, and that
task will figure out the rest of what's needed to ultimately send the
correct message to the workers.

The older code is being left in place for now so things don't break all
over the place while I verify that the logic is working.

Next up will be to create a simpler version of the generate.py python
code that just does a single transformation and returns a message with
the same taskid (uuid) it was given. That should complete the loop for
finally making transformations run independently. With that in place,
we can start to decouple other code that used to need to run alongside
a transformation chain, such as download preparation and importing
column data.