turn a name into a dataset while respecting access
This encapsulates some more complex logic that used to be spread
throughout the system. Basically, a user is trying to register a
transformation, referencing some input datasets by name. If they
include the organization name ('org:dataset'), simple! Just make sure
they're not using an org they don't really have access to. If they
don't give an organization, then we assume the non-prefixed dataset is
in the same organization as the output dataset. This is captured in this
function by nearDataset (which doesn't actually have to be an output
dataset). It's used to figure out which organization to create under if
the organization is not given explicitly, while still making sure the
user can access everything. If we can't find the dataset or the user
can't access it, we return null.