create a RegisterTask subclass
RegisterTask handles the status of registering transformation code.
Since we can now enforce permissions on the server, the work that
register_transformation.py has to do is vastly simplified. Now, it only
has to pick up the names used in dataset_input calls. It resolves them
(or not) when the register task finishes.