convert import_csv to the new task model
Unfortunately, there's no great way to break this commit up. There was
a lot of untangling to do and quite a few obvious problems that would
have been made temporarily worse if I'd broken it down more. Here's a
short laundry list of what was done.

- Lots of code was removed. This felt so good.
- ImportCSVTask was created in the same way the others have been. I'm
sensing that these may warrent their own subfolder before long, just to
keep things tidy. As well, there's a bit of boilerplate developing that
I could probably eliminate with the odd extra Task function. For now,
it's still much better than before, so I may leave that for another
time.
- Previous to this, when column info was computed for a generated
dataset, tags were not automatically set like they are when you import
a raw dataset. There were other weird inconsistencies as well.
Happily, moving more logic back to the server domain objects made
solving these inconsistencies pretty straightforward. So now we grab
tags properly in both cases, yay! At some point, we may be able to do
away with the duplication of import functionality and just attach an
import task after every transform task, but let's keep from going too
crazy with Tasks right off.
- import_csv.py probably had the most removed. It still had some
database code as well. The really nice thing is that these python
scripts are becoming pretty simple and to the point. With
import_csv.py, it's a lot easier to see now that it really just needs
to create a normalized csv, a sample csv, and try and figure out some
column names/types. That was a lot harder to follow before.
- There was a removeExisting flag before (and still some now inert
references to get rid of). This was always kind of a weird kludge. The
only time we set it to true before was when we were dealing with CSV
import errors. Otherwise, we just assumed you never wanted to remove
columns. If, say, you were uploading new data to a dataset with changed
columns, this could create some weird results. The real reason it
ever needed to exist was that we weren't quite as comfortable with
Cypher queries and maintaining the graph. It's a lot easier now to
simply look at the new column names, remove any existing columns that
don't match up, then merge any new ones in. So we're getting rid of the
weird need to think about removeExisting. If we wanted to get really
fancy, now that we have a Task node where we can store intermediate
information involved with tasks, we could even delay setting new column
info until a user confirmed it.
- Finally, removed the this.importTask setting. Let's just use
overriding the function. That makes the logic a little easier to follow.