use new tasks to prepare download
We'll still probably want to get away from automatically running tasks
(and maybe even waiting for downloads, opting instead to return a
response that they're not yet ready). But this at least allows us to do
away with some of the weirder callback stuff that was happening before.
And we no longer need to do a specific download task on the queue! Yay!
This is because we're now using the same mechanism to prep the download
as when the user generates the dataset.