can't enforce access rules after a deletion
We were returning the deleted object when doing a delete mutation in
GraphQL, which is sort of convention. But when we're also trying to
enforce access rights that depend on looking something up in the
database, that doesn't really work. So now we'll just return true for
success (or throw an error like we normally do).