fix broken migration
Migrations still need database connections. It does make less and less
sense for them to use the worker container to operate, though. In the
meantime, this restores what this migration needs, and it still works,
though we've been a pretty long time from having data that would need
the migration.