Parameterize docker-compose image tag
This enables us to specify the tag we'd like to use when building and
running docker containers, which will be especially useful when we're
automatically building and pushing them.

Note that we also add a .env file, which specifies a default of
'latest', meaning this change shouldn't actually break anything if you
don't go out of your way to set the TAG environment variable.