intermediate module for redis setup
This simply creates 2 clients, one async and one not, and exports them.
Ideally, we'd only have to create one client and just wrap it for the
async version, but the async version seems to do a little too much to
the object it decorates, which makes that one incompatible with the
connect-redis package being used for the session store.