add access to redis in the repl
It's really useful to be able to play with redis in the repl for
debugging, deciding on keys, etc. Only the async version is being
imported. Outside of the necessary callback version for connect-redis,
the async version is way more natural to work with under modern
async/await style nodejs code.