manage column visibility through domain class w/ redis
This moves the weird, hacky in-memory column visibility code to the
Column class. Furthermore, the Column class uses redis instead of
relying on memory (which will go away with every restart). Now column
visibility will stick in the cache. It's still not important enough to
manage through the graph database, but it's kind of nice that it's not
as easy to disappear.