fix null proptype warning
I discovered this while I was verifying something else. I'm used to
using false-y values in if conditions on their own, without checking if
they're actually null or undefined. I believe this is idiomatic for
javascript and many other languages. However, the React components are
doing some extra proptype checking, and if they get a null, they'll
complain (even though it still works because it's false-y). I wavered
as to whether to just wrap the entire bit of logic in a Boolean or
explicitly check for null. I think checking for null is the better path
here. Though it feels less idiomatic, it more clearly shows what's
going on, as we never expect inputTransformation to be something that
is actually true/false.